+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## neptune-tensorflow-keras 2.2.1
+
+### Changes
+- Added `log_model_summary` parameter to specify if the model summary should be logged ([#60](https://github.com/neptune-ai/neptune-tensorflow-keras/pull/60))
+
+### Fixes
+- Handled cases when model summary or visualization cannot be logged ([#60](https://github.com/neptune-ai/neptune-tensorflow-keras/pull/60))
+
 ## neptune-tensorflow-keras 2.2.0
 
 ### Fixes